# src/apps/api/app/main.py
import re
import uuid

import structlog
from fastapi import FastAPI, Request, Response, status
from fastapi.middleware.cors import CORSMiddleware
from fastapi.middleware.gzip import GZipMiddleware
from fastapi.responses import JSONResponse
from slowapi import _rate_limit_exceeded_handler
from slowapi.errors import RateLimitExceeded
from sqlalchemy.exc import DBAPIError

from app.agents._runtime import (
    AiBudgetExceededError,
    AiCostControlPolicyError,
    CostControlUnavailableError,
    ProviderOutcomeUnknownError,
)
from app.config import settings
from app.limiter import limiter
from app.routes import (
    admin,
    admin_build_tasks,
    admin_creator_memory,
    admin_creator_style,
    admin_edit_feedback,
    admin_generative,
    admin_jobs,
    admin_kria,
    admin_music,
    admin_plan_items,
    admin_review,
    admin_sound_effects,
    auth,
    clips,
    content_plans,
    creation_threads,
    creator_agent,
    creator_workspace,
    dev_qa_storage,
    generative_jobs,
    kria_runtime,
    landing,
    manual_drafts,
    me,
    memory,
    music,
    music_jobs,
    personas,
    plan_items,
    presigned,
    sound_effects,
    template_jobs,
    templates,
    tiktok,
    uploads,
    waitlist,
)

log = structlog.get_logger()

app = FastAPI(title="Kria API", version="0.1.0")
app.state.limiter = limiter
app.add_exception_handler(RateLimitExceeded, _rate_limit_exceeded_handler)

_CORS_ALLOW_ORIGIN_REGEX = r"https://nova-.*-emirerbens-projects\.vercel\.app"
# usekria.com apex 308-redirects to www.usekria.com (Vercel domain config), so the
# origin every real browser sends is the `www` form — it must be allowlisted
# alongside the apex, not instead of it.
_CANONICAL_WEB_ORIGIN = "https://www.usekria.com"
_APEX_WEB_ORIGIN = "https://usekria.com"
_LEGACY_WEB_ORIGIN = "https://nova-video.vercel.app"
_CODE_ALLOWED_ORIGINS = {_CANONICAL_WEB_ORIGIN, _APEX_WEB_ORIGIN, _LEGACY_WEB_ORIGIN}


def _allowed_origins() -> list[str]:
    return sorted(set(settings.allowed_origins) | _CODE_ALLOWED_ORIGINS)


app.add_middleware(
    CORSMiddleware,
    allow_origins=_allowed_origins(),
    allow_origin_regex=_CORS_ALLOW_ORIGIN_REGEX,
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Job-status polls ship 100KB+ of JSON every 2s per client and nothing between
# uvicorn and the Vercel proxy compresses (the proxy re-encodes toward the
# browser). minimum_size spares tiny health/status bodies.
app.add_middleware(GZipMiddleware, minimum_size=1024)


def _safe_trace_id(value: str | None) -> str | None:
    if value and re.fullmatch(r"[A-Za-z0-9._:-]{1,128}", value):
        return value
    return None


@app.middleware("http")
async def request_correlation(request: Request, call_next):  # noqa: ANN001
    """Give every HTTP hop a unique request ID and preserve the batch ID."""
    request_id = _safe_trace_id(request.headers.get("x-request-id")) or uuid.uuid4().hex
    correlation_id = _safe_trace_id(request.headers.get("x-correlation-id")) or request_id
    request.state.request_id = request_id
    request.state.correlation_id = correlation_id
    structlog.contextvars.clear_contextvars()
    structlog.contextvars.bind_contextvars(
        request_id=request.state.request_id,
        correlation_id=request.state.correlation_id,
    )
    try:
        response = await call_next(request)
    finally:
        structlog.contextvars.clear_contextvars()
    response.headers["X-Request-Id"] = request.state.request_id
    response.headers["X-Correlation-Id"] = request.state.correlation_id
    return response


def _cors_headers_for(request: Request) -> dict[str, str]:
    """Return Access-Control-* headers for the given request's Origin.

    Starlette installs `@app.exception_handler(Exception)` at the
    ServerErrorMiddleware level, which sits OUTSIDE CORSMiddleware in the
    middleware stack (encode/starlette#1175). Responses produced by this
    handler therefore bypass CORSMiddleware entirely on the way out, so we
    have to mirror the allow-origin logic here. Without this, the browser
    sees a 500 with no Access-Control-Allow-Origin and surfaces a
    TypeError: Failed to fetch instead of letting the frontend read the
    status code.
    """
    origin = request.headers.get("origin")
    if not origin:
        return {}
    allowed = origin in _allowed_origins() or bool(re.fullmatch(_CORS_ALLOW_ORIGIN_REGEX, origin))
    if not allowed:
        return {}
    return {
        "Access-Control-Allow-Origin": origin,
        "Access-Control-Allow-Credentials": "true",
        "Vary": "Origin",
    }


#: PostgreSQL transient-concurrency SQLSTATEs.  ``40P01`` is deadlock_detected:
#: two transactions took the same row locks in opposite order and the server
#: aborted one of them.  ``40001`` is serialization_failure.  Both mean "your
#: transaction lost a race", not "the request was invalid" -- the same call
#: succeeds on retry, so it must not surface as a 500.
_RETRYABLE_SQLSTATES = frozenset({"40P01", "40001"})


def _transient_sqlstate(exc: BaseException) -> str | None:
    """Return the SQLSTATE if ``exc`` wraps a retryable serialization error."""

    if not isinstance(exc, DBAPIError):
        return None
    sqlstate = getattr(exc.orig, "sqlstate", None) or getattr(exc.orig, "pgcode", None)
    return sqlstate if sqlstate in _RETRYABLE_SQLSTATES else None


@app.exception_handler(AiBudgetExceededError)
async def ai_budget_exhausted_handler(request: Request, exc: AiBudgetExceededError) -> JSONResponse:
    """Return one stable, cache-aware contract for every paid-call breaker."""

    request_id = (
        getattr(request.state, "request_id", None)
        or _safe_trace_id(request.headers.get("x-request-id"))
        or uuid.uuid4().hex
    )
    correlation_id = (
        getattr(request.state, "correlation_id", None)
        or _safe_trace_id(request.headers.get("x-correlation-id"))
        or request_id
    )
    log.warning(
        "ai_budget_exhausted",
        path=request.url.path,
        scope=exc.scope,
        reason=exc.reason,
        reset_at=exc.reset_at,
        request_id=request_id,
    )
    return JSONResponse(
        status_code=429,
        content={
            "detail": "AI budget exhausted for this scope.",
            "code": "ai_budget_exhausted",
            "scope": exc.scope,
            "reset_at": exc.reset_at,
            "cached_behavior_available": exc.cached_behavior_available,
            "reason": exc.reason,
            "retryable": False,
            "request_id": request_id,
            "correlation_id": correlation_id,
        },
        headers={
            **_cors_headers_for(request),
            "X-Request-Id": request_id,
            "X-Correlation-Id": correlation_id,
        },
    )


@app.exception_handler(AiCostControlPolicyError)
async def ai_cost_policy_handler(request: Request, exc: AiCostControlPolicyError) -> JSONResponse:
    """Return a correction-required contract, never a fake budget reset."""

    request_id = getattr(request.state, "request_id", None) or uuid.uuid4().hex
    correlation_id = getattr(request.state, "correlation_id", None) or request_id
    return JSONResponse(
        status_code=exc.status_code,
        content={
            "detail": "Paid AI attribution or approval is invalid.",
            "code": "ai_cost_control_policy_rejected",
            "scope": exc.scope,
            "reason": exc.reason,
            "retryable": False,
            "request_id": request_id,
            "correlation_id": correlation_id,
        },
        headers={
            **_cors_headers_for(request),
            "X-Request-Id": request_id,
            "X-Correlation-Id": correlation_id,
        },
    )


@app.exception_handler(CostControlUnavailableError)
async def ai_cost_control_unavailable_handler(
    request: Request, exc: CostControlUnavailableError
) -> JSONResponse:
    """Fail closed with an explicit control-plane outage contract."""

    request_id = getattr(request.state, "request_id", None) or uuid.uuid4().hex
    correlation_id = getattr(request.state, "correlation_id", None) or request_id
    return JSONResponse(
        status_code=503,
        content={
            "detail": "AI cost controls are unavailable; no provider call was made.",
            "code": "ai_cost_control_unavailable",
            "retryable": True,
            "request_id": request_id,
            "correlation_id": correlation_id,
        },
        headers={
            **_cors_headers_for(request),
            "X-Request-Id": request_id,
            "X-Correlation-Id": correlation_id,
        },
    )


@app.exception_handler(ProviderOutcomeUnknownError)
async def ai_provider_outcome_unknown_handler(
    request: Request, exc: ProviderOutcomeUnknownError
) -> JSONResponse:
    """Do not invite a second paid request while the first outcome is unknown."""

    request_id = getattr(request.state, "request_id", None) or uuid.uuid4().hex
    correlation_id = getattr(request.state, "correlation_id", None) or request_id
    log.warning(
        "ai_provider_outcome_unknown",
        path=request.url.path,
        request_id=request_id,
        error=str(exc)[:300],
    )
    return JSONResponse(
        status_code=503,
        content={
            "detail": "The AI request may still be processing. Do not retry this operation.",
            "code": "ai_provider_outcome_unknown",
            "retryable": False,
            "request_id": request_id,
            "correlation_id": correlation_id,
        },
        headers={
            **_cors_headers_for(request),
            "X-Request-Id": request_id,
            "X-Correlation-Id": correlation_id,
        },
    )


@app.exception_handler(DBAPIError)
async def transient_db_conflict_handler(request: Request, exc: DBAPIError) -> JSONResponse:
    """Surface deadlock/serialization aborts as a retryable 409, never a 500."""

    sqlstate = _transient_sqlstate(exc)
    if sqlstate is None:
        return await unhandled_exception_handler(request, exc)
    request_id = (
        getattr(request.state, "request_id", None)
        or _safe_trace_id(request.headers.get("x-request-id"))
        or uuid.uuid4().hex
    )
    correlation_id = (
        getattr(request.state, "correlation_id", None)
        or _safe_trace_id(request.headers.get("x-correlation-id"))
        or request_id
    )
    # Deadlocks are expected under concurrency and are self-healing, but a
    # sustained rate means a lock-order inversion slipped past the static
    # guard in tests/routes/test_lock_order.py -- so log it as a warning with
    # the SQLSTATE rather than dropping it silently.
    log.warning(
        "transient_db_conflict",
        path=request.url.path,
        method=request.method,
        sqlstate=sqlstate,
        request_id=request_id,
        correlation_id=correlation_id,
    )
    return JSONResponse(
        status_code=409,
        content={
            "detail": "That project was being updated at the same time. Try again.",
            "code": "concurrent_update",
            "retryable": True,
            "request_id": request_id,
            "correlation_id": correlation_id,
        },
        headers={
            **_cors_headers_for(request),
            "Retry-After": "1",
            "X-Request-Id": request_id,
            "X-Correlation-Id": correlation_id,
        },
    )


@app.exception_handler(Exception)
async def unhandled_exception_handler(request: Request, exc: Exception) -> JSONResponse:
    request_id = (
        getattr(request.state, "request_id", None)
        or _safe_trace_id(request.headers.get("x-request-id"))
        or uuid.uuid4().hex
    )
    correlation_id = (
        getattr(request.state, "correlation_id", None)
        or _safe_trace_id(request.headers.get("x-correlation-id"))
        or request_id
    )
    log.exception(
        "unhandled_exception",
        path=request.url.path,
        method=request.method,
        request_id=request_id,
        correlation_id=correlation_id,
    )
    return JSONResponse(
        status_code=500,
        content={
            "detail": "Kria couldn't complete that request. Retry in a moment.",
            "code": "internal_error",
            "retryable": True,
            "request_id": request_id,
            "correlation_id": correlation_id,
        },
        headers={
            **_cors_headers_for(request),
            "X-Request-Id": request_id,
            "X-Correlation-Id": correlation_id,
        },
    )


app.include_router(auth.router, prefix="/auth", tags=["auth"])
app.include_router(uploads.router, prefix="/uploads", tags=["uploads"])
app.include_router(admin.router, prefix="/admin", tags=["admin"])
app.include_router(admin_jobs.router, prefix="/admin/jobs", tags=["admin-jobs"])
app.include_router(admin_kria.router, prefix="/admin/kria", tags=["admin-kria"])
app.include_router(admin_plan_items.router, prefix="/admin/plan-items", tags=["admin-plan-items"])
app.include_router(admin_music.router, prefix="/admin/music-tracks", tags=["admin-music"])
app.include_router(
    admin_sound_effects.router, prefix="/admin/sound-effects", tags=["admin-sound-effects"]
)
app.include_router(admin_generative.router, prefix="/admin/generative", tags=["admin-generative"])
app.include_router(
    admin_creator_memory.router,
    prefix="/admin/creator-memory",
    tags=["admin-creator-memory"],
)
app.include_router(
    admin_build_tasks.router, prefix="/admin/build-tasks", tags=["admin-build-tasks"]
)
app.include_router(admin_review.router, prefix="/admin/review", tags=["admin-review"])
app.include_router(
    admin_edit_feedback.router,
    prefix="/admin/edit-feedback",
    tags=["admin-edit-feedback"],
)
app.include_router(
    admin_creator_style.router,
    prefix="/admin/creator-style-assignments",
    tags=["admin-creator-style"],
)
app.include_router(template_jobs.router, prefix="/template-jobs", tags=["template-jobs"])
app.include_router(music.router, prefix="/music-tracks", tags=["music"])
app.include_router(sound_effects.router, prefix="/sound-effects", tags=["sound-effects"])
app.include_router(music_jobs.router, prefix="/music-jobs", tags=["music-jobs"])
app.include_router(generative_jobs.router, prefix="/generative-jobs", tags=["generative-jobs"])
app.include_router(personas.router, prefix="/personas", tags=["personas"])
app.include_router(content_plans.router, prefix="/content-plans", tags=["content-plans"])
app.include_router(creation_threads.router, prefix="/creation-threads", tags=["creation-threads"])
app.include_router(kria_runtime.router, prefix="/creation-threads", tags=["kria-runtime-v2"])
app.include_router(plan_items.router, prefix="/plan-items", tags=["plan-items"])
app.include_router(creator_agent.router, prefix="/plan-items", tags=["creator-agent"])
app.include_router(creator_workspace.router, prefix="/content-plans", tags=["creator-workspace"])
app.include_router(manual_drafts.router, prefix="/plan-items", tags=["plan-items"])
app.include_router(me.router, prefix="/me", tags=["me"])
app.include_router(memory.router, prefix="/me", tags=["memory"])
app.include_router(memory.direction_router, tags=["direction-overrides"])
app.include_router(presigned.router, prefix="/presigned-urls", tags=["presigned"])
app.include_router(clips.router, prefix="/clips", tags=["clips"])
app.include_router(templates.router, prefix="/templates", tags=["templates"])
app.include_router(tiktok.router, prefix="/tiktok", tags=["tiktok"])
app.include_router(waitlist.router, tags=["waitlist"])
app.include_router(landing.router, prefix="/landing-clips", tags=["landing"])
app.include_router(dev_qa_storage.router, prefix="/dev-qa/storage", tags=["dev-qa"])


@app.get("/health")
async def health() -> dict:
    return {"status": "ok"}


@app.get("/health/beat")
async def health_beat(response: Response) -> dict:
    """Reports whether Celery Beat is alive, via a heartbeat any
    Beat-scheduled task writes on success (app/services/beat_heartbeat.py).

    Pinged from OUTSIDE this app by .github/workflows/beat-health.yml
    (15-min cron) — the only kind of check immune to Beat's own death,
    since any check that is itself Beat-scheduled shares Beat's exact
    blind spot. Matters
    specifically because RENDER_AUTOSTOP_ENABLED makes Beat's health a
    prerequisite for the render worker ever restarting on a missed
    wake-hook call (see agents/DECISIONS.md).

    Returns 503 (not 200) when unhealthy — most external uptime monitors
    alert on a non-2xx status alone without needing to parse the body, so
    the status code IS the actionable signal, not just the JSON payload.
    """
    from app.services.beat_heartbeat import beat_heartbeat_status  # noqa: PLC0415

    healthy, age_seconds = beat_heartbeat_status()
    if not healthy:
        response.status_code = status.HTTP_503_SERVICE_UNAVAILABLE
    return {
        "status": "ok" if healthy else "stale",
        "age_seconds": age_seconds,
    }