# src/apps/ios/Kria/Generated/openapi.yaml
{
  "components": {
    "schemas": {
...
      "CreateBody": {
        "additionalProperties": false,
        "properties": {
          "client_event_id": {
            "default": null,
            "maxLength": 160,
            "nullable": true,
            "title": "Client Event Id",
            "type": "string"
          },
          "message": {
            "default": null,
            "maxLength": 4000,
            "nullable": true,
            "title": "Message",
            "type": "string"
          },
          "runtime_version": {
            "default": 1,
            "enum": [
              1,
              2
            ],
            "title": "Runtime Version",
            "type": "integer"
          }
        },
        "title": "CreateBody",
        "type": "object"
      },
      "CreationThreadOut": {
        "properties": {
          "active_creator_agent_session_id": {
            "default": null,
            "nullable": true,
            "title": "Active Creator Agent Session Id",
            "type": "string"
          },
          "active_job_id": {
            "default": null,
            "nullable": true,
            "title": "Active Job Id",
            "type": "string"
          },
          "active_plan_item_id": {
            "default": null,
            "nullable": true,
            "title": "Active Plan Item Id",
            "type": "string"
          },
          "content_plan_id": {
            "default": null,
            "nullable": true,
            "title": "Content Plan Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "creator_agent": {
            "additionalProperties": true,
            "default": null,
            "nullable": true,
            "title": "Creator Agent",
            "type": "object"
          },
          "direction_receipt": {
            "additionalProperties": true,
            "default": null,
            "nullable": true,
            "title": "Direction Receipt",
            "type": "object"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventOut"
            },
            "title": "Events",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "job": {
            "additionalProperties": true,
            "default": null,
            "nullable": true,
            "title": "Job",
            "type": "object"
          },
          "media_capabilities": {
            "additionalProperties": true,
            "default": null,
            "nullable": true,
            "title": "Media Capabilities",
            "type": "object"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "runtime_version": {
            "default": 1,
            "enum": [
              1,
              2
            ],
            "title": "Runtime Version",
            "type": "integer"
          },
          "speech_cleanup": {
            "additionalProperties": true,
            "default": null,
            "nullable": true,
            "title": "Speech Cleanup",
            "type": "object"
          },
          "state": {
            "additionalProperties": true,
            "title": "State",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "revision",
          "state",
          "events",
          "created_at",
          "updated_at"
        ],
        "title": "CreationThreadOut",
        "type": "object"
      },
...
      "EventOut": {
        "properties": {
          "content": {
            "default": null,
            "nullable": true,
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "default": null,
            "nullable": true,
            "title": "Payload",
            "type": "object"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "sequence",
          "revision",
          "role",
          "event_type",
          "created_at"
        ],
        "title": "EventOut",
        "type": "object"
      },
...
      "PersonaResponse": {
        "properties": {
          "error_detail": {
            "nullable": true,
            "title": "Error Detail",
            "type": "string"
          },
          "generation_started_at": {
            "default": null,
            "format": "date-time",
            "nullable": true,
            "title": "Generation Started At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "idea_seeds": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Idea Seeds",
            "type": "array"
          },
          "persona": {
            "additionalProperties": true,
            "nullable": true,
            "title": "Persona",
            "type": "object"
          },
          "persona_status": {
            "title": "Persona Status",
            "type": "string"
          },
          "questionnaire": {
            "additionalProperties": true,
            "nullable": true,
            "title": "Questionnaire",
            "type": "object"
          },
          "tiktok_profile": {
            "additionalProperties": true,
            "default": null,
            "nullable": true,
            "title": "Tiktok Profile",
            "type": "object"
          },
          "tiktok_style_analysis_available": {
            "default": false,
            "title": "Tiktok Style Analysis Available",
            "type": "boolean"
          },
          "tiktok_style_analysis_failed_at": {
            "default": null,
            "format": "date-time",
            "nullable": true,
            "title": "Tiktok Style Analysis Failed At",
            "type": "string"
          },
          "tiktok_style_analysis_failure_reason": {
            "default": null,
            "nullable": true,
            "title": "Tiktok Style Analysis Failure Reason",
            "type": "string"
          },
          "tiktok_style_analysis_status": {
            "default": "idle",
            "enum": [
              "idle",
              "running",
              "ready",
              "failed"
            ],
            "title": "Tiktok Style Analysis Status",
            "type": "string"
          },
          "tiktok_style_analyzed_at": {
            "default": null,
            "format": "date-time",
            "nullable": true,
            "title": "Tiktok Style Analyzed At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "persona_status",
          "questionnaire",
          "persona",
          "error_detail"
        ],
        "title": "PersonaResponse",
        "type": "object"
      },
      "QuestionnaireBody": {
        "properties": {
          "hobbies": {
            "default": "",
            "title": "Hobbies",
            "type": "string"
          },
          "location": {
            "default": "",
            "title": "Location",
            "type": "string"
          },
          "passions": {
            "default": "",
            "title": "Passions",
            "type": "string"
          },
          "school": {
            "default": "",
            "title": "School",
            "type": "string"
          },
          "social": {
            "default": "",
            "title": "Social",
            "type": "string"
          },
          "tiktok_handle": {
            "default": "",
            "title": "Tiktok Handle",
            "type": "string"
          },
          "travels": {
            "default": "",
            "title": "Travels",
            "type": "string"
          },
          "work": {
            "default": "",
            "title": "Work",
            "type": "string"
          }
        },
        "title": "QuestionnaireBody",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "mobileBearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Checked native subset; server drafts and render jobs remain authoritative.",
    "title": "Kria Mobile API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/creation-threads": {
      "get": {
        "operationId": "listCreationThreads",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CreationThreadOut"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Success"
          }
        },
        "security": [
          {
            "mobileBearer": []
          }
        ]
      },
      "post": {
        "operationId": "createCreationThread",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreationThreadOut"
                }
              }
            },
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Revision or identity conflict"
          },
          "422": {
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "mobileBearer": []
          }
        ]
      }
    },
...
    "/personas": {
      "get": {
        "operationId": "getPersona",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaResponse"
                }
              }
            },
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Revision or identity conflict"
          },
          "422": {
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "mobileBearer": []
          }
        ]
      },
      "post": {
        "operationId": "createPersona",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionnaireBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaResponse"
                }
              }
            },
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Revision or identity conflict"
          },
          "422": {
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "mobileBearer": []
          }
        ]
      }
    }
  },
  "servers": [
    {
      "url": "https://nova-video.fly.dev"
    }
  ]
}